pyproject.toml
[tool.poetry]
name = "books-demo"
version = "0.0.0"
description = ""
authors = ["monkeysees <email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = {extras = ["dotenv"], version = "^2.7.4"}
mypy = "^1.10.0"
motor = "^3.4.0"
pydantic-settings = "^2.3.3"
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"
pillow = "^10.3.0"
types-aiofiles = "^23.2.0.20240403"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
